* Middleware mod refactoring * Performance improvements
