+++ README.md
@@ -53,12 +53,12 @@ RAM: 31686MiB
 ### Results
 ```
 Running 10s test @ http://127.0.0.1:7878/hello
-  1 threads and 200 connections
+  4 threads and 500 connections
   Thread Stats   Avg      Stdev     Max   +/- Stdev
-    Latency   578.90us  206.77us   5.77ms   79.81%
-    Req/Sec   184.72k     9.54k  200.74k    77.00%
-  1837693 requests in 10.08s, 206.80MB read
-Requests/sec: 182380.80
-Transfer/sec:     20.52MB
+    Latency     1.39ms    1.05ms  18.45ms   81.47%
+    Req/Sec     89.69k    18.07k  126.91k   57.50%
+  3575551 requests in 10.07s, 395.55MB read
+Requests/sec: 355053.82
+Transfer/sec: 39.28MB
 ```
